+++ app/controllers/projects_controller.rb
@@ -19,7 +19,7 @@ class ProjectsController < ApplicationController
     stratum_keys = params.keys.select{|key| key =~ /^stratum_[\d]*$/}
     values = params.each.select{|key, value| stratum_keys.include?(key)}.collect{|k,v| v}.compact
 
-    if @assignment = @project.create_randomization!(params[:subject_code], values)
+    if @assignment = @project.create_randomization!(params[:subject_code], values, current_user)
       redirect_to [@project, @assignment], notice: "Subject successfully randomized to <b>#{@assignment.treatment_arm}</b>.".html_safe
     else
       render action: 'randomize_subject'